correcion realizada fr base de datos

+++ database/migrations/2023_12_18_030346_create_category_lesson_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('category_lesson', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('category_id');
-            $table->unsignedInteger('lesson_id');
+            $table->unsignedBigInteger('category_id');
+            $table->unsignedBigInteger('lesson_id');
             $table->timestamps();
         });
     }
